modelStore sharedInstance로 리팩토링
