+++ ios/Mukkaebie/Mukkaebie/AppDelegate.swift
@@ -63,6 +63,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     func applicationDidBecomeActive(_ application: UIApplication) {
+        NetworkStore.getStoreList { (storeList) in
+            Store.sharedInstance.allStores = storeList
+        }
+        
         // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
     }
