+++ ios/Mukkaebie/Mukkaebie/NetworkStore.swift
@@ -12,24 +12,24 @@ import Alamofire
 
 class NetworkStore {
     
-    private let url = URLpath.getURL()
+    private static let url = URLpath.getURL()
     
-    func getStoreList() {
+    static func getStoreList(callback: @escaping (_ storeList: [ModelStores]) -> Void) {
+        var storeList = [ModelStores]()
+
         Alamofire.request("\(url)stores").responseJSON { (response) in
             if let response = response.result.value as? [[String:Any]] {
-                var storeList = [ModelStores]()
                 for item in response {
                     let store = ModelStores(JSON: item)
                     storeList.append(store!)
                 }
-
-                NotificationCenter.default.post(name: NSNotification.Name(rawValue: "getStoreList"), object: nil, userInfo: ["storeList":storeList])
+                callback(storeList)
             }
         }
     }
     
     func getStoreList(sellerId: Int) {
-        let encodedString = ("\(url)stores/\(sellerId)").addingPercentEncoding(withAllowedCharacters: CharacterSet.urlFragmentAllowed)
+        let encodedString = ("\(NetworkStore.url)stores/\(sellerId)").addingPercentEncoding(withAllowedCharacters: CharacterSet.urlFragmentAllowed)
         
         Alamofire.request(URL(string: encodedString!)!).responseJSON { (response) in
             if let response = response.result.value as? [[String:Any]] {
@@ -46,21 +46,4 @@ class NetworkStore {
         }
     }
     
-    func getStoreList(category: String) {
-        let encodedString = ("\(url)stores/bycategory/"+category).addingPercentEncoding(withAllowedCharacters: CharacterSet.urlFragmentAllowed)
-        
-        Alamofire.request(URL(string: encodedString!)!).responseJSON { (response) in
-            if let response = response.result.value as? [[String:Any]] {
-                var storeList = [ModelStores]()
-                for item in response {
-                    let store = ModelStores(JSON: item)
-                    storeList.append(store!)
-                }
-                
-                NotificationCenter.default.post(name: NSNotification.Name(rawValue: "getStoreList"), object: nil, userInfo: ["storeList":storeList])
-            } else {
-                print(response.result)
-            }
-        }
-    }
 }
