+++ ios/Mukkaebie/Mukkaebie/NetworkUser.swift
@@ -28,30 +28,3 @@ class NetworkUser {
     }
     
 }
-
-
-/*
-static func getUser(callback: @escaping (_ result: Bool, _ modelUser: ModelUser) -> Void) {
-    Alamofire.request("\(url)/api/v1/user/login").responseJSON { (response) in
-        var modelUser = ModelUser()
-        
-        let res = JSON(data: response.data!)
-        let result = res["result"].boolValue
-        if let user = res["user"].dictionary {
-            if let id = user["_id"]?.stringValue,
-                let email = user["email"]?.stringValue,
-                let isKakaoImage = user["isKakaoImage"]?.boolValue,
-                let nickname = user["nickname"]?.stringValue,
-                let bookmark = user["bookmark"]?.arrayValue.map({ $0.stringValue }) {
-                var imageURL = String()
-                if let userImageURL = user["imageURL"]?.stringValue {
-                    imageURL = userImageURL
-                }
-                modelUser = ModelUser(id: id, isKakaoImage: isKakaoImage, email: email, nickname: nickname, bookmark: bookmark, profileImageURL: imageURL
-                )
-            }
-        }
-        callback(result, modelUser)
-    }
-}
-*/
